Clean magic item descriptions (#415)
* restore * to Deck of Many Things

Some items are denoted with asterisks in the source material, which was
not included here.

* add formatting to magic items

In the source material, effect listings (eg. for individual cards in the
Deck of Many Things) are in bold italic. Updated markdown formatting to
match.

* rm sentient magic items description

The description of sentient magic items got lumped in with Wings of
Flying somehow.

* remove potion of healing from Equipment.json

It's a magic item.