fix: Edit content with rule changes and typographical fixes from SRD 5.1  (#529)
* fix: Align subclass feature text with SRD 5.1

SRD 5.1 removed references to subclasses not documented in the SRD,
as well as references to "the end of the class description".

SRD 5.1 also removed references to "Archfey" and "Great Old One"
from warlock class features, including the Pact Boon examples.

* fix: Align bag of tricks, spell scroll item descriptions to SRD 5.1

Lowercase "hit points" for the bag of tricks creature.

Update the spell scroll item rules to SRD 5.1, which specifies that
only material components are exempted, changes the usage time
required from an action to the spell's casting time, and specifies
that the scroll is not destroyed if spellcasting is interrupted.

Fix typos and missing "spell scroll" references in the per-level
spell scroll objects.

* fix: Specify Ambush behavior limit for doppleganger

SRD 5.1 specifies that the doppleganger's Ambush ability applies
only on the first round of combat.

* fix: Remove line about drow alignment from Elf racial description

SRD 5.1 removes a line stating that most drow are evil.

* fix: Replace DM references with GM

SRD 5.1 replaces all references to DM, a protected term, with GM,
the open term.

* fix: Punctuation errors and appendix references

SRD 5.1 renamed the appendices to add initialized references to
sources. For example, "appendix A" referring to conditions is
renamed to "appendix PH-A".

Remove extraneous spaces from hyphens and dashes, and convert
some hyphens to dashes (`--`).