fix: Add subclass improvements to level features (#836)
## What does this do?

This PR adds features for each of the subclass improvement stages, and
adds them to the appropriate level. The intent is that this should make
the API data reflect the content of the class table in the SRD document.

The Barbarian Primal Path improvements were added first to ensure that
the result is correct before replicating the effort for the remaining
classes.


![image](https://github.com/user-attachments/assets/6a6a6729-54f9-4ee9-81ce-6596d6909674)

Work now progresses on the remaining classes.

- [x] Barbarian
- [x] Bard
- [x] Cleric
- [x] Druid
- [x] Fighter
- [x] Monk
- [x] Paladin
- [x] Ranger
- [x] Rogue
- [x] Sorcerer
- [x] Warlock
- [x] Wizard

~~Once complete, the intention is to convert this PR from Draft to a
full PR.~~
As all classes are now complete, this PR has been elevated to full PR
for further review.

## How was it tested?

Built locally (non-Docker), and resultant tables inspected via MongoDB
Compass.


![image](https://github.com/user-attachments/assets/164f82a5-416f-476b-90b9-e2b54810dc9f)

Local tests run and pass successfully.

## Is there a Github issue this is resolving?

Although briefly discussed on Discord, a GitHub Issue has not been
raised at this time.

## Did you update the docs in the API? Please link an associated PR if
applicable.

No - I don't believe a document update is required, but please correct
me if I am wrong.

## Here's a fun image for your troubles


![image](https://github.com/user-attachments/assets/b772c8ac-5407-4d4a-bdd4-7a817eb29219)